backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Product Initiative Manager backend"
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = ">=0.135.2,<0.136.0"
uvicorn = ">=0.42.0,<0.43.0"
PyGithub = ">=2.0"
pydantic-settings = ">=2.0.0"
python-multipart = ">=0.0.9"

[tool.poetry.group.test.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.24"
httpx = ">=0.27"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
